requirements.txt
agent-framework>=1.0.0b251120
fastapi>=0.115.0
uvicorn>=0.32.0
azure-identity>=1.19.0
python-dotenv>=1.0.0
pydantic>=2.0.0
